fix: missing nifi processors
