+++ ldi-nifi/ldi-nifi-common/src/test/java/be/vlaanderen/informatievlaanderen/ldes/ldi/validators/RDFLanguageValidatorTest.java
@@ -1,5 +1,6 @@
 package be.vlaanderen.informatievlaanderen.ldes.ldi.validators;
 
+import be.vlaanderen.informatievlaanderen.ldes.ldi.processors.validators.RDFLanguageValidator;
 import org.apache.nifi.components.ValidationContext;
 import org.apache.nifi.components.ValidationResult;
 import org.junit.jupiter.api.Test;
