+++ ldi-nifi/ldi-nifi-processors/create-version-object-processor/src/main/resources/META-INF/services/org.apache.nifi.processor.Processor
@@ -1,0 +1,1 @@
+be.vlaanderen.informatievlaanderen.ldes.ldi.processors.CreateVersionObjectProcessor
