+++ ldi-nifi/ldi-nifi-processors/ldes-client-processor/src/main/resources/META-INF/services/org.apache.nifi.processor.Processor
@@ -1,0 +1,1 @@
+be.vlaanderen.informatievlaanderen.ldes.ldi.processors.LdesClient
