+++ ldi-nifi/ldi-nifi-processors/sparql-interactions-processor/src/main/resources/META-INF/services/org.apache.nifi.processor.Processor
@@ -1,0 +1,2 @@
+be.vlaanderen.informatievlaanderen.ldes.ldi.processors.SparqlConstructProcessor
+be.vlaanderen.informatievlaanderen.ldes.ldi.processors.SparqlSelectProcessor
