--- ldi-nifi/ldi-nifi-processors/sparql-interactions-processor/src/test/java/be/vlaanderen/informatievlaanderen/ldes/ldi/TestUtils.java
@@ -1,5 +1,0 @@
-package be.vlaanderen.informatievlaanderen.ldes.ldi;
-
-public class TestUtils {
-
-}
