+++ ldi-nifi/ldi-nifi-processors/version-materialisation-processor/src/main/resources/META-INF/services/org.apache.nifi.processor.Processor
@@ -1,0 +1,1 @@
+be.vlaanderen.informatievlaanderen.ldes.ldi.processors.VersionMaterialiseProcessor
